feat(geometry/euclidean/oriented_angle): oriented angles between points
Define the oriented angle between three points, notation `∡`, in terms
of that between two vectors, and set up some corresponding basic API
(including relating it to unoriented angles between three points).
For angles between points, the choice of orientation is implicit, via
use of the `module.oriented` typeclass to specify a preferred
orientation.

---

This provides some initial basic API that doesn't depend on anything
else not yet merged.  Further lemmas are ready that depend on #16243
or #16278, and further lemmas are planned that depend on #16184 or