feat(archive/birthday): improve birthday problem (#16528)
This adds a measure-theoretic interpretation of the birthday problem; not yet the one with independent uniform random variables, but some form of a step in that direction.

It also removes a misleading statement from the YAML to make it clearer that `card_embedding_eq` is not all the work done on the Birthday problem on mathlib. cc @digama0



Co-authored-by: Eric Rodriguez <37984851+ericrbg@users.noreply.github.com>