feat(group_theory/schreier): The size of the commutator subgroup is bounded in terms of the index of the center and the number of commutators (#16679)
This PR proves that the size of the commutator subgroup is bounded in terms of the index of the center and the number of commutators. The proof uses Schreier's lemma and the transfer homomorphism.

I included lots of comments since the proof is rather technical. But please let me know if I went overboard.

Ultimately, this is building up to a bound on the size of the commutator subgroup just in terms of the number of commutators.



Co-authored-by: Thomas Browning <tb65536@users.noreply.github.com>