feat(linear_algebra/affine_space/affine_subspace): spans of two points
Add lemmas about the `vector_span` and `affine_span of two points
(analogous to `span_singleton` lemmas for `submodule`).

---

Although I previously noted in review of #10632 that I expected such
lemmas to be of relevance to proving Ceva's theorem in appropriate
generality, my immediate motivation for this PR is various other
geometrical uses.