feat(linear_algebra/affine_space/independent): affine independence of all but one point (#16815)
Add the lemma that, in an affine space for a module over a division ring, if all but one point of a family are affinely independent, and that point does not lie in the affine span of the other points, the family is affinely independent.  Thus, deduce lemmas that three points are affinely independent if two are distinct and the third does not lie in an affine subspace containing those two.