feat(ring_theory): specialize Chinese Remainder Theorem to two ideals
The current arbitrarily-indexed version of CRT is surprisingly hard to apply to `R / (I ⊓ J)` inline, so here it is extracted into its own theorem.

I added an import of the `fin_cases` tactic to `ring_theory.ideal.quotient`; I believe this is lightweight enough and certainly it wouldn't cause a big jump in the transitive imports.