feat(analysis/special_functions/trigonometric/angle): `to_real` lemmas
Add a series of lemmas mainly about `real.angle.to_real` and twice
angles, concluding with conditions for an angle and twice that angle
to have the same sign (which is of use for proving that the base
angles of an isosceles triangle are acute).

---

There may well be scope for golfing the proof of
`sign_two_nsmul_eq_sign_iff`, but even if various of the other lemmas
end up no longer used in that proof after such golfing, I think they
are all sensible lemmas for mathlib independent of use in that proof.